// src/styles/themes/types.ts
export interface FlavorColors {
  rosewater: string;
  flamingo: string;
  pink: string;
  mauve: string;
  red: string;
  maroon: string;
  peach: string;
  yellow: string;
  green: string;
  teal: string;
  sky: string;
  sapphire: string;
  blue: string;
  lavender: string;
  text: string;
  subtext1: string;
  subtext0: string;
  overlay2: string;
  overlay1: string;
  overlay0: string;
  surface2: string;
  surface1: string;
  surface0: string;
  base: string;
  mantle: string;
  crust: string;
}

export interface ThemeDefinition {
  id: string;
  label: string;
  dark: boolean;
  colors: FlavorColors;
}

export interface CustomTheme {
  id: string;
  name: string;
  dark: boolean;
  colors: FlavorColors;
}